Stacking

+++ Assets/Scripts/Bullet.cs
@@ -8,7 +8,7 @@ public class Bullet : MonoBehaviour
 
     [Header("Attributes")]
     [SerializeField] private float bulletSpeed = 5f;
-    [SerializeField] private int bulletDamage = 1;
+    [SerializeField] private float bulletDamage = 1;
     [SerializeField] private StatusEffect statusEffect;
     [SerializeField] private float tickRate;
     [SerializeField] private float bulletLifetime = 5f;
